Move ICanSpecifyAssemblies from Execution to Configuration.Assemblies.
