+++ Source/Bifrost.Specs/Bootstrap/Assemblies/for_AssemblySpecifiers/AnotherTestSpecifier.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using Bifrost.Configuration.Assemblies;
-using Bifrost.Execution;
 using Moq;
 
 namespace Bifrost.Specs.Bootstrap.Assemblies.for_AssemblySpecifiers
