+++ Source/Bifrost.Specs/Bootstrap/Assemblies/for_AssemblySpecifiers/TestSpecifier.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using Bifrost.Configuration.Assemblies;
-using Bifrost.Execution;
 using Moq;
 
 namespace Bifrost.Specs.Bootstrap.Assemblies.for_AssemblySpecifiers
