+++ Source/Bifrost/AssemblySpecifier.cs
@@ -17,7 +17,6 @@
 //
 #endregion
 using Bifrost.Configuration.Assemblies;
-using Bifrost.Execution;
 
 namespace Bifrost
 {
